pyproject.toml
[project]
name = "pucly"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "django-quill-editor (>=0.1.42,<0.2.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "channels (>=4.2.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
